+++ Cargo.toml
@@ -13,25 +13,23 @@ description = """PagerDuty API from Rust."""
 
 
 [features]
-default = ["ureq"]
-async = ["futures", "futures-util", "reqwest"]
+default = ["async"]
+
+# Enable reqwest's blocking client on sync
+sync = ["reqwest/blocking"]
+
+# Need futures for Async
+async = []
 
 [dependencies]
 url = "2.2.0"
 time = {version = "0.2.23", features = ["std", "serde"]}
 serde = "1.0.118"
-serde_json = {version = "1.0.61" }
-
-# Optional - Only when using sync
-ureq = {version = "1.5.4", optional = true }
-
-# Optional - only enabled through the "async" feature
-futures = { version = "0.3.8", optional = true }
-futures-util = { version = "0.3.8", optional = true }
-reqwest = { version = "0.10.10", features = ["json"], optional = true }
+reqwest = { version = "0.11.0", features = ["json"]}
 
 [dev-dependencies]
 assert_matches = "1.4.0"
 pretty_assertions = "0.6.1"
-rand = "0.8.0"
-tokio = { version = "1.0.1", features = ["rt", "macros"], optional = true }
+rand = "0.8.1"
+tokio = { version = "1.0.1", features = ["rt", "macros"]}
+serde_json = "1.0.61"
