+++ src/eventsv2sync.rs
@@ -1,53 +1,47 @@
 use crate::private_types::*;
 use crate::types::*;
 
+use reqwest::blocking::Client;
+use reqwest::header::{
+    HeaderMap, HeaderValue, InvalidHeaderValue, CONTENT_ENCODING, CONTENT_TYPE, USER_AGENT,
+};
 use serde::Serialize;
 use std::convert::From;
 use std::error::Error;
 use std::fmt::{Display, Formatter, Result as FmtResult};
-use std::time::Duration;
-
-const CONTENT_TYPE: &str = "content-type";
-const USER_AGENT: &str = "user-agent";
-const CONTENT_ENCODING: &str = "content-encoding";
 
 const CONTENT_ENCODING_IDENTITY: &str = "identity";
 const CONTENT_TYPE_JSON: &str = "application/json";
 
 #[derive(Debug)]
 pub enum EventsV2Error {
-    SerdeJsonError(serde_json::Error),
+    ReqwestError(reqwest::Error),
+    InvalidHeaderValue(InvalidHeaderValue),
 
-    /// A synthetic error from ureq library (basically not an HTTP error of any kind.)
-    SyntheticUreqError(Option<ureq::Error>),
-
-    /// Any Http Error
-    HttpError(u16),
-
-    /// We expected PagerDuty to respond with 202 Accepted - this indicates we got anything but 202
-    /// could be 1xx or 3xx codes as well.
-    NotAccepted(u16),
+    //https://developer.pagerduty.com/docs/events-api-v2/overview/#api-response-codes--retry-logic
+    HttpNotAccepted(u16), // NOT 4xx, 5xx or 200 (we expect 202). Contains HTTP response code.
+    HttpError(u16),       // A legit error (4xx or 5xx). Contains HTTP response code.
 }
 
 impl Error for EventsV2Error {}
 impl Display for EventsV2Error {
     fn fmt(&self, f: &mut Formatter) -> FmtResult {
         match self {
-            Self::SerdeJsonError(e) => write!(f, "SerdeJsonError: {}", e),
-            Self::SyntheticUreqError(oe) => match oe {
-                Some(e) => write!(f, "SyntheticUreqError: {}", e),
-                None => write!(f, "SyntheticUreqError (no details)"),
-            },
-            Self::HttpError(c) => write!(f, "HttpError with Status Code {}", c),
-            Self::NotAccepted(c) => {
-                write!(f, "Http Status Code was other than 202 (Accepted): {}", c)
-            }
+            Self::ReqwestError(e) => write!(f, "RequestError: {}", e),
+            Self::InvalidHeaderValue(e) => write!(f, "InvalidHeaderValue: {}", e),
+            Self::HttpNotAccepted(e) => write!(f, "HttpNotAccepted: {}", e),
+            Self::HttpError(e) => write!(f, "HttpError: {}", e),
         }
     }
 }
-impl From<serde_json::Error> for EventsV2Error {
-    fn from(err: serde_json::Error) -> Self {
-        Self::SerdeJsonError(err)
+impl From<reqwest::Error> for EventsV2Error {
+    fn from(err: reqwest::Error) -> Self {
+        Self::ReqwestError(err)
+    }
+}
+impl From<InvalidHeaderValue> for EventsV2Error {
+    fn from(err: InvalidHeaderValue) -> Self {
+        Self::InvalidHeaderValue(err)
     }
 }
 
@@ -57,7 +51,7 @@ pub type EventsV2Result = Result<(), EventsV2Error>;
 pub struct EventsV2 {
     /// The integration/routing key for a generated PagerDuty service
     integration_key: String,
-    user_agent: Option<String>,
+    client: Client,
 }
 
 impl EventsV2 {
@@ -65,9 +59,21 @@ impl EventsV2 {
         integration_key: String,
         user_agent: Option<String>,
     ) -> Result<EventsV2, EventsV2Error> {
+        let mut headers = HeaderMap::new();
+        headers.insert(CONTENT_TYPE, HeaderValue::from_str(CONTENT_TYPE_JSON)?);
+        headers.insert(
+            CONTENT_ENCODING,
+            HeaderValue::from_str(CONTENT_ENCODING_IDENTITY)?,
+        );
+        if let Some(ua) = user_agent {
+            headers.insert(USER_AGENT, HeaderValue::from_str(ua.as_str())?);
+        }
+
+        let client = Client::builder().default_headers(headers).build()?;
+
         Ok(EventsV2 {
             integration_key,
-            user_agent,
+            client,
         })
     }
 
@@ -111,296 +117,12 @@ impl EventsV2 {
 
     // Make this part Async in the future
     fn do_post<T: Serialize>(&self, url: &str, content: T) -> EventsV2Result {
-        let jsonstr = serde_json::to_string(&content)?;
+        let res = self.client.post(url).json(&content).send()?;
 
-        let mut request = ureq::post(url);
-
-        request
-            .set(CONTENT_TYPE, CONTENT_TYPE_JSON)
-            .set(CONTENT_ENCODING, CONTENT_ENCODING_IDENTITY)
-            // 300 seconds should be plenty to post to pagerduty
-            .timeout(Duration::from_secs(300));
-
-        if let Some(ua) = &self.user_agent {
-            request.set(USER_AGENT, ua);
-        };
-
-        let resp = request.send_string(jsonstr.as_str());
-
-        if resp.synthetic() {
-            return Err(EventsV2Error::SyntheticUreqError(
-                resp.into_synthetic_error(),
-            ));
+        match res.status().as_u16() {
+            202 => Ok(()),
+            e if e < 400 => Err(EventsV2Error::HttpNotAccepted(e)),
+            e => Err(EventsV2Error::HttpError(e)),
         }
-
-        if resp.error() {
-            return Err(EventsV2Error::HttpError(resp.status()));
-        }
-
-        if resp.status() != 202 {
-            return Err(EventsV2Error::NotAccepted(resp.status()));
-        }
-
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-    use time::OffsetDateTime;
-
-    #[derive(Serialize)]
-    pub struct SerializableTest {
-        some_field: String,
-        another_field: usize,
-    }
-
-    #[test]
-    fn serialize_change() {
-        // With everything optional
-        let c = Change {
-            payload: ChangePayload {
-                summary: "Hello".to_owned(),
-                source: Some("hostname".to_owned()),
-                timestamp: OffsetDateTime::from_unix_timestamp_nanos(2000071804323000000),
-                custom_details: Some(SerializableTest {
-                    some_field: "Serialize this!".to_owned(),
-                    another_field: 34,
-                }),
-            },
-            links: Some(vec![Link {
-                href: "https://polyverse.com".to_owned(),
-                text: Some("Polyverse homepage".to_owned()),
-            }]),
-        };
-
-        let cr = serde_json::to_string(&c);
-        assert!(cr.is_ok());
-        assert_eq!(cr.unwrap(), "{\"payload\":{\"summary\":\"Hello\",\"timestamp\":\"2033-05-18T23:30:04.323000000Z\",\"source\":\"hostname\",\"custom_details\":{\"some_field\":\"Serialize this!\",\"another_field\":34}},\"links\":[{\"href\":\"https://polyverse.com\",\"text\":\"Polyverse homepage\"}]}");
-
-        // With nothing optional
-        let c = Change::<()> {
-            payload: ChangePayload {
-                summary: "Hello".to_owned(),
-                timestamp: OffsetDateTime::from_unix_timestamp_nanos(2000071804323000000),
-                source: None,
-                custom_details: None,
-            },
-            links: None,
-        };
-
-        let cr = serde_json::to_string(&c);
-        assert!(cr.is_ok());
-        assert_eq!(
-            cr.unwrap(),
-            "{\"payload\":{\"summary\":\"Hello\",\"timestamp\":\"2033-05-18T23:30:04.323000000Z\"}}"
-        );
-    }
-
-    #[test]
-    fn serialize_sendable_change() {
-        // With everything optional
-        let c = SendableChange {
-            routing_key: "routingkey".to_owned(),
-            payload: ChangePayload {
-                summary: "Hello".to_owned(),
-                source: Some("hostname".to_owned()),
-                timestamp: OffsetDateTime::from_unix_timestamp_nanos(2000071804323000000),
-                custom_details: Some(SerializableTest {
-                    some_field: "Serialize this!".to_owned(),
-                    another_field: 34,
-                }),
-            },
-            links: Some(vec![Link {
-                href: "https://polyverse.com".to_owned(),
-                text: Some("Polyverse homepage".to_owned()),
-            }]),
-        };
-
-        let cr = serde_json::to_string(&c);
-        assert!(cr.is_ok());
-        assert_eq!(cr.unwrap(), "{\"routing_key\":\"routingkey\",\"payload\":{\"summary\":\"Hello\",\"timestamp\":\"2033-05-18T23:30:04.323000000Z\",\"source\":\"hostname\",\"custom_details\":{\"some_field\":\"Serialize this!\",\"another_field\":34}},\"links\":[{\"href\":\"https://polyverse.com\",\"text\":\"Polyverse homepage\"}]}");
-
-        // With nothing optional
-        let c = SendableChange::<()> {
-            routing_key: "routingkey".to_owned(),
-            payload: ChangePayload {
-                summary: "Hello".to_owned(),
-                timestamp: OffsetDateTime::from_unix_timestamp_nanos(2000071804323000000),
-                source: None,
-                custom_details: None,
-            },
-            links: None,
-        };
-
-        let cr = serde_json::to_string(&c);
-        assert!(cr.is_ok());
-        assert_eq!(cr.unwrap(), "{\"routing_key\":\"routingkey\",\"payload\":{\"summary\":\"Hello\",\"timestamp\":\"2033-05-18T23:30:04.323000000Z\"}}");
-    }
-
-    #[test]
-    fn serialize_alert_trigger() {
-        // With everything optional
-        let a = AlertTrigger {
-            payload: AlertTriggerPayload {
-                summary: "Hello".to_owned(),
-                source: "hostname".to_owned(),
-                timestamp: Some(OffsetDateTime::from_unix_timestamp_nanos(
-                    2000071804323000000,
-                )),
-                severity: Severity::Info,
-                component: Some("postgres".to_owned()),
-                group: Some("prod-datapipe".to_owned()),
-                class: Some("deploy".to_owned()),
-                custom_details: Some(SerializableTest {
-                    some_field: "Serialize this!".to_owned(),
-                    another_field: 34,
-                }),
-            },
-            dedup_key: Some("dedupkey1".to_owned()),
-            images: Some(vec![Image {
-                src: "https://polyverse.com/static/img/SplashPageIMG/polyverse_blue.png".to_owned(),
-                href: Some("https://polyverse.com".to_owned()),
-                alt: Some("The Polyverse Logo".to_owned()),
-            }]),
-            links: Some(vec![Link {
-                href: "https://polyverse.com".to_owned(),
-                text: Some("Polyverse homepage".to_owned()),
-            }]),
-            client: Some("Zerotect".to_owned()),
-            client_url: Some("https://github.com/polyverse/zerotect".to_owned()),
-        };
-
-        let ar = serde_json::to_string(&a);
-        assert!(ar.is_ok());
-        assert_eq!(ar.unwrap(), "{\"payload\":{\"severity\":\"info\",\"summary\":\"Hello\",\"source\":\"hostname\",\"timestamp\":\"2033-05-18T23:30:04.323000000Z\",\"component\":\"postgres\",\"group\":\"prod-datapipe\",\"class\":\"deploy\",\"custom_details\":{\"some_field\":\"Serialize this!\",\"another_field\":34}},\"dedup_key\":\"dedupkey1\",\"images\":[{\"src\":\"https://polyverse.com/static/img/SplashPageIMG/polyverse_blue.png\",\"href\":\"https://polyverse.com\",\"alt\":\"The Polyverse Logo\"}],\"links\":[{\"href\":\"https://polyverse.com\",\"text\":\"Polyverse homepage\"}],\"client\":\"Zerotect\",\"client_url\":\"https://github.com/polyverse/zerotect\"}");
-
-        // With nothing optional
-        let a = AlertTrigger::<()> {
-            payload: AlertTriggerPayload {
-                summary: "Hello".to_owned(),
-                source: "hostname".to_owned(),
-                timestamp: None,
-                severity: Severity::Info,
-                component: None,
-                group: None,
-                class: None,
-                custom_details: None,
-            },
-            dedup_key: None,
-            images: None,
-            links: None,
-            client: None,
-            client_url: None,
-        };
-
-        let ar = serde_json::to_string(&a);
-        assert!(ar.is_ok());
-        assert_eq!(
-            ar.unwrap(),
-            "{\"payload\":{\"severity\":\"info\",\"summary\":\"Hello\",\"source\":\"hostname\"}}"
-        );
-    }
-
-    #[test]
-    fn serialize_sendable_alert_trigger() {
-        // With everything optional
-        let a = SendableAlertTrigger {
-            routing_key: "routingkey".to_owned(),
-            event_action: Action::Trigger,
-            payload: AlertTriggerPayload {
-                summary: "Hello".to_owned(),
-                source: "hostname".to_owned(),
-                timestamp: Some(OffsetDateTime::from_unix_timestamp_nanos(
-                    2000071804323000000,
-                )),
-                severity: Severity::Info,
-                component: Some("postgres".to_owned()),
-                group: Some("prod-datapipe".to_owned()),
-                class: Some("deploy".to_owned()),
-                custom_details: Some(SerializableTest {
-                    some_field: "Serialize this!".to_owned(),
-                    another_field: 34,
-                }),
-            },
-            dedup_key: Some("dedupkey1".to_owned()),
-            images: Some(vec![Image {
-                src: "https://polyverse.com/static/img/SplashPageIMG/polyverse_blue.png".to_owned(),
-                href: Some("https://polyverse.com".to_owned()),
-                alt: Some("The Polyverse Logo".to_owned()),
-            }]),
-            links: Some(vec![Link {
-                href: "https://polyverse.com".to_owned(),
-                text: Some("Polyverse homepage".to_owned()),
-            }]),
-            client: Some("Zerotect".to_owned()),
-            client_url: Some("https://github.com/polyverse/zerotect".to_owned()),
-        };
-
-        let ar = serde_json::to_string(&a);
-        assert!(ar.is_ok());
-        assert_eq!(ar.unwrap(), "{\"routing_key\":\"routingkey\",\"payload\":{\"severity\":\"info\",\"summary\":\"Hello\",\"source\":\"hostname\",\"timestamp\":\"2033-05-18T23:30:04.323000000Z\",\"component\":\"postgres\",\"group\":\"prod-datapipe\",\"class\":\"deploy\",\"custom_details\":{\"some_field\":\"Serialize this!\",\"another_field\":34}},\"dedup_key\":\"dedupkey1\",\"images\":[{\"src\":\"https://polyverse.com/static/img/SplashPageIMG/polyverse_blue.png\",\"href\":\"https://polyverse.com\",\"alt\":\"The Polyverse Logo\"}],\"links\":[{\"href\":\"https://polyverse.com\",\"text\":\"Polyverse homepage\"}],\"event_action\":\"trigger\",\"client\":\"Zerotect\",\"client_url\":\"https://github.com/polyverse/zerotect\"}");
-
-        // With nothing optional
-        let a = SendableAlertTrigger::<()> {
-            routing_key: "routingkey".to_owned(),
-            event_action: Action::Trigger,
-            payload: AlertTriggerPayload {
-                summary: "Hello".to_owned(),
-                source: "hostname".to_owned(),
-                timestamp: None,
-                severity: Severity::Info,
-                component: None,
-                group: None,
-                class: None,
-                custom_details: None,
-            },
-            dedup_key: None,
-            images: None,
-            links: None,
-            client: None,
-            client_url: None,
-        };
-
-        let ar = serde_json::to_string(&a);
-        assert!(ar.is_ok());
-        assert_eq!(ar.unwrap(), "{\"routing_key\":\"routingkey\",\"payload\":{\"severity\":\"info\",\"summary\":\"Hello\",\"source\":\"hostname\"},\"event_action\":\"trigger\"}");
-    }
-
-    #[test]
-    fn serialize_alert_acknowledge() {
-        let a = AlertAcknowledge {
-            dedup_key: "dedupkeyacknowledge".to_owned(),
-        };
-
-        let ar = serde_json::to_string(&a);
-        assert!(ar.is_ok());
-        assert_eq!(ar.unwrap(), "{\"dedup_key\":\"dedupkeyacknowledge\"}");
-    }
-
-    #[test]
-    fn serialize_alert_resolve() {
-        let a = AlertResolve {
-            dedup_key: "dedupkeyacknowledge".to_owned(),
-        };
-
-        let ar = serde_json::to_string(&a);
-        assert!(ar.is_ok());
-        assert_eq!(ar.unwrap(), "{\"dedup_key\":\"dedupkeyacknowledge\"}");
-    }
-
-    #[test]
-    fn serialize_sendable_alert_followup() {
-        let ss = SendableAlertFollowup {
-            dedup_key: "DedupkeyFollowup".to_owned(),
-            routing_key: "routingkey".to_owned(),
-            event_action: Action::Resolve,
-        };
-
-        let ssr = serde_json::to_string(&ss);
-        assert!(ssr.is_ok());
-        assert_eq!(ssr.unwrap(), "{\"routing_key\":\"routingkey\",\"dedup_key\":\"DedupkeyFollowup\",\"event_action\":\"resolve\"}");
     }
 }
